Clarify role of vertex data in examples
This adds a simple comment that clarifies how the data in the
vertex_format struct relates to the shader.
I struggled a bit before I understood this, so maybe it will be helpful
for others to have that piece of information in the examples (at least
until more comprehensive documentation exists).