Implement base indexes
This removes `IndexCount`, as it will never be distinct from `VertexCount`,
and cleans up the `Slice` representation to factor out common fields with
names, because it was becoming unmanageable. Changes the `DrawCommand`
slightly to take a buffer handle when binding the index buffer, because the
size of the buffer is needed in the fallback for when base indexes are not
supported. This code path should be very rare.

First half of #355