Make Batch trait object-safe
Breaking change!

By removing the associated error type the batch trait becomes object-safe. The
advantage is, that it allows run-time dispatch of batches, as long as the
backing Resources implementation is known at compile-time.

The downside is, of course, that the error type is less generic. This is deemed
less useful, as it only conveys information in case of errors and the additional
Error::Other enum variant allows for any kind of String to be passed as an
error.

As an alternative one could implement an additional trait `DynamicBatch<R>`
implemented for all `B: Batch<R>`. This can also be done in user code. However,
I think that the possibility to dynamically dispatch batches warrants this
change.