[ll] First API draft for pipeline and shader handling
The new API is based on the already existing concept of pipeline state objects. In this approach the PSO are additionally coupled with a RenderPass and the shader entry points.
Instead of loading single shaders we now have libraries to fit the SPIR-V module approach from Vulkan and the metallibs from Metal. On other APIs these libraries are manually built.