Hilarous experiments with GL state caching.
Basically, adds a filter that tries not to emit redundant calls into
the CommandBuffer.

This would ideally be done at a *higher* level than the CommandBuffer,
but this is simpler to implement to get off the ground quickly, and
does actually seem to work half decently.  This is about where the
easy part of it ends; dealing with attributes and such has more sneaky
interactions that have to be dealt with.