Adding more caching to GL backend.
Would resolve (a large part of) issue #1198 .

It's quite a brute-force solution, and I am deeply uncertain that it tracks everything correctly, but I tried to err
on the side of safety, allowing redundant state-setting calls rather than getting rid of them. It is also incomplete:
there is some stuff, such as stencil buffers, that it doesn't try to track (because I don't really know how). However,
it appears to work; I don't see any errors in any of the examples, and apitrace output for the performance
example is about equivalent to the gl example (which is good!).