Add display_fb field to gl CommandBuffer
Adds a new public field to the CommandBuffer structure in the gfx_device_gl backend. This allows to use an existing Framebuffer object as the drawing destination instead of the Framebuffer with name 0. Useful on iOS where no default frame buffer exists and drawing can only occur on system provided renderbuffers.