Fix Metal backend's color handling
Two issues were causing distored colors when using the Metal backend:

1. RGBA8 was mapped to BGRA8 when rendering, to accomodate CAMetalLayer's requirement that
it only use BGRA ordering for 8-bit backbuffer formats. This however created issues when handling textures,
so instead we expose this difference to the user. It should only require them to choose a backbuffer format
based on the backend used, which is already needed for Vulkan.

2. CAMetalLayer appears to interpret the buffer as sRGB whether you want it to or not. To accomodate this we
use the backbuffer color format (B8_G8_R8_A8, Srgb) in gfx_app when using Metal, so writes to the backbuffer
are transformed properly. This may actually be an error on our side, because we do some hacky things with
regards to backbuffer management, but until those are fixed this will at least produce correct results.