Fixed rendering and presentation
Metal's SwapChain implementation now presents the buffers directly, without any copying
by setting the IOSurface backing the Metal texture as the contents of a CALayer. This hands
our renderbuffer straight to the compositor, just as the native swapchain in CAMetalLayer does.

Additionally, the trianglell example now displays almost correctly when using Metal; it now samples
the texture properly instead of displaying the texture coordinates as colors. However blend mode
interpretation is still not implemented, so the background of the rectangle is white, not transparent.