Separate Buffer/Encoder for RenderPass Encoders
To avoid needing to include traits with lifetime parameters in
the GraphicsCommandBuffer trait, we split the RenderPass*Encoder trait
into RenderPass*Buffer (a backend-implemented trait with no lifetime
parameters) and a struct RenderPass*Encoder containing the necessary
lifetime parameters and associated references.

These structs proxy all their calls to the Buffer trait implementations
by passing a reference to the whole struct that includes the buffer as
a member, as well as the lifetime-bound references. As a result the
backend implementations have access to all the data they had previously
without having to resort to lifetime parameters on the trait or
unsafety.