[ll] Refactor command pools following the command buffer concept of differing between raw and subpass
After the implementations of the backends we saw that the all of them made use of macros for implementing command pools.
Introducing the concept of RawCommandPools, similar to RawCommandBuffers, allows us to remove the need of macros and simplifies backend implementations.

Current state only updated the GL backend (breaks other backends and examples).