Remove stride parameter on buffer creation
Stride was only needed for dx12 to use it for vertex buffer binding. The PSO stride and the stride on buffer creation needed to be equal, causing issues on the user side. Caching the vertex buffers and use the strides from the active PSO fixes this. Also, helps for better vulkan portability.