metal: Add debug capture scopes
Because of the way Swapchains are implemented, Xcode's frame capture
gets confused on where the frame starts and ends (which is the default
capture scope).

This commit creates a capture scope that tries to mimic the usual
behaviour. Enabled only when debug_assertations is on or with feature
metal_default_capture_scope.