metal: Hack to fix the first frame of a swapchain
The freeing up of the drawable now checks if there is anything to free
from the has_swapchain boolean.

The boolean is totally a hack, and might have some edge-cases where it
is wrong, but this happens to fix warnings and red-frame flashes
(present of not drawn frame) from the quad example, with the resizing
still working.