Fixed copying wrong color buffer to tex in single-buffered contexts
See #2310

In double buffered contexts, GL_BACK is available to copy the color buffer from.
However, in single buffered contexts this is not true, so calling
copy_texture_to_buffer_raw when using a single buffered context resulted in a GL
error.

This commit contains a temporary fix which just glGets the current draw buffer and
uses that, which seems to fix the issue.