Add a texture format filter.
This commit adds a structure called `info::TextureFormatFilter` and changes the
code in `info::query_all` so that, when detecting OpenGL ES it can select a
set of valid texture property combinations.

This provides `PhysicalDevice::image_format_properties` with a way of more
accurately filtering out formats that aren't valid for textures and whose
use would lead to a crash.